+++ Sources/TUSKit/Tasks/IdentifiableTask.swift
@@ -1,0 +1,12 @@
+//
+//  IdentifiableTask.swift
+//  
+//
+//  Created by Elvirion Antersijn on 26/02/2022.
+//
+
+import Foundation
+
+protocol IdentifiableTask: ScheduledTask {
+    var id: UUID { get }
+}