+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Added
 - ChunkSize argument to TUSClient initializer.
+- Add cancel single task.
 
 # 3.0.0
 - Rewrite of TUSKit